console quiz crud

+++ Web Quiz Engine/task/src/engine/QuizService.java
@@ -2,5 +2,5 @@ package engine;
 
 public interface QuizService {
     public Quiz createQuiz(Quiz quiz);
-    public Quiz findQuizById(int id);
+    public Quiz findQuizById(Long id);
 }
